Rename getHierarchizedData to getTreeFromArray
